requirements.txt
absl-py==2.1.0
altair==5.2.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dm-tree==0.1.8
exceptiongroup==1.2.0
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.10.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
redis==5.0.3
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.31.1
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
